+++ src/components/Header.tsx
@@ -1,31 +1,7 @@
-import { useRecoilState, useRecoilValue } from 'recoil';
-
-import HeaderItem from './HeaderItem';
-import { userState } from '../recoil/recoil';
 import styled from 'styled-components';
 
 const Header = () => {
-  const [users, setUsers] = useRecoilState(userState);
-
-  const handleClickChangeUser = (user: number) => {};
-
-  return (
-    <HeaderWrapper>
-      {users.map((it) => (
-        <HeaderItem key={it.user_id} {...it} />
-      ))}
-    </HeaderWrapper>
-  );
+  return;
 };
-
-const HeaderWrapper = styled.div`
-  flex-grow: 0.5;
-
-  padding-left: 1.5rem;
-
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-`;
 
 export default Header;
