+++ src/components/Navigator.tsx
@@ -1,0 +1,5 @@
+const Navigator = () => {
+  return;
+};
+
+export default Navigator;
