+++ src/pages/Setting.tsx
@@ -1,0 +1,19 @@
+import styled from 'styled-components';
+
+const Setting = () => {
+  return <Container></Container>;
+};
+
+const Container = styled.div`
+  width: 25rem;
+  height: 40rem;
+
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+
+  border-radius: 20px;
+  box-shadow: 1px 1px 10px lightgray;
+`;
+
+export default Setting;
